Add files via upload
The code is to detect the disk failure by using GAN ( Generative Adversarial Networks). SMART features are multivariate time series. And the number of the failed hard disks is very small compared to the number of normal hard disks, in which case supervised learning is not effective.
As in a typical GAN, the generator (G) generates fake samples from a specific latent space and passes that to the discriminator(D) which tries to distinguish fake from real. Based on the outputs of D , the system will update parameters of D and G, so that the discriminator will be trained to be as sensitive as possible to assign correct labels to both real and fake samples, while the generator will be trained as smart as possible to fool the discriminator (assigning real labels to fake samples). After sufficient rounds of iterations, the generator will have captures the hidden distributions of the training sequences and that could generate realistic samples. And D will be an intuitive tool for anomaly detection(disk failure detection).